fix: add settlement to admin api

+++ src/main/java/com/example/repick/domain/clothingSales/dto/GetClothingSalesProduct.java
@@ -9,7 +9,7 @@ public record GetClothingSalesProduct(
         @Schema(description = "등급") String grade,
         @Schema(description = "판매기간") String salesPeriod,
         @Schema(description = "판매 금액") Long salesPrice,
-        @Schema(description = "정산금") Integer settlementPrice,
+        @Schema(description = "정산금") Long settlementPrice,
         @Schema(description = "수수료") Integer fee,
         @Schema(description = "신청일") String requestDate,
         @Schema(description = "반송 여부") Boolean isReturned
